DAOS-2190 rsvc: Move get/set/list attributes into RSVC module
The code for getting, setting and listing attributes is duplicated
in both Pool and Container services. This patch consolidates common
functionality into the recently created Replicated Service module.
