+++ src/rsvc/SConscript
@@ -9,7 +9,7 @@ def scons():
     denv = env.Clone()
 
     # ds_rsvc
-    ds_rsvc = daos_build.library(denv, 'rsvc', ['srv.c'])
+    ds_rsvc = daos_build.library(denv, 'rsvc', ['srv.c', 'srv_common.c'])
     denv.Install('$PREFIX/lib/daos_srv', ds_rsvc)
 
 if __name__ == "SCons.Script":
